log startup messages to stderr

+++ cbf/bin/build.sh
@@ -24,7 +24,7 @@ source "$(readlink -f "$( dirname "${BASH_SOURCE[0]}" )")/init.libraries"
 
 # build our container
 timer.measureCmd "$NAME" 'cbf.buildContainer' "$NAME" "$DEBUG_TRACE" "$TZ"
-rm "$(crf.propertiesFile)"
-rm "$(cbf.propertiesFile)"
+[ -w "$(crf.propertiesFile)" ] && rm "$(crf.propertiesFile)"
+[ -w "$(cbf.propertiesFile)" ] && rm "$(cbf.propertiesFile)"
 echo ''
 echo ''
